Split tests into separate classes and fixed method names that violated coding standard
