+++ src/main/java/mistermusik/logic/Command.java
@@ -1,6 +1,5 @@
 package mistermusik.logic;
 
-import mistermusik.Main;
 import mistermusik.commons.Contact;
 import mistermusik.commons.events.eventtypes.Event;
 import mistermusik.commons.events.eventtypes.eventsubclasses.assessmentsubclasses.Exam;
@@ -244,7 +243,7 @@ public class Command {
                 String[] checklistCommand = splitChecklist[0].split(" ");
                 int eventIndex = Integer.parseInt(checklistCommand[1]) - 1;
 //                if(!((events.getEvent(eventIndex).getType()=='P') || (events.getEvent(eventIndex).getType()=='L'))) {
-//                    ui.noSuchEvent();
+//                    ui.printNoSuchEvent();
 //                    return;
 //                }
                 if (checklistCommand.length == 3) {
@@ -276,9 +275,9 @@ public class Command {
                     }
                 }
             } catch (IndexOutOfBoundsException ne) {
-                ui.noSuchEvent();
+                ui.printNoSuchEvent();
             } catch (NumberFormatException numE) {
-                ui.notAnInteger();
+                ui.printNotAnInteger();
             }
         }
     }
@@ -317,7 +316,7 @@ public class Command {
      */
     private void editEvent(EventList events, UI ui) {
         if (continuation.isEmpty()) {
-            ui.eventDescriptionEmpty();
+            ui.printEventDescriptionEmpty();
         } else {
             String[] splitInfo = continuation.split("/");
             int eventIndex = Integer.parseInt(splitInfo[0]) - 1;
@@ -333,14 +332,14 @@ public class Command {
      */
     private void showOrSetBudget(EventList events, UI ui) {
         if (continuation.isEmpty()) {
-            ui.budgetCommandWrongFormat();
+            ui.printBudgetCommandWrongFormat();
         } else if (continuation.substring(0, 3).equals("set")) { //set new budget
             try {
                 int newBudget = Integer.parseInt(continuation.substring(4));
                 events.getBudgeting().setBudget(newBudget);
-                UI.budgetSet(newBudget);
+                UI.printBudgetSet(newBudget);
             } catch (NumberFormatException e) {
-                ui.notAnInteger();
+                ui.printNotAnInteger();
             }
         } else { //show budget for given month
             String monthAndYear = continuation;
@@ -356,7 +355,7 @@ public class Command {
     //@@author
     private void searchEvents(EventList events, UI ui) {
         if (continuation.isEmpty()) {
-            ui.eventDescriptionEmpty();
+            ui.printEventDescriptionEmpty();
         } else {
             String searchKeyWords = continuation;
             String foundEvent = "";
@@ -426,7 +425,7 @@ public class Command {
     //@@author Ryan-Wong-Ren-Wei
     private void addNewEvent(EventList events, UI ui, char eventType) {
         if (continuation.isEmpty()) {
-            ui.eventDescriptionEmpty();
+            ui.printEventDescriptionEmpty();
         } else {
             try {
                 EntryForEvent entryForEvent = new EntryForEvent().invoke(); //separate all info into relevant details
@@ -437,7 +436,7 @@ public class Command {
                 if (entryForEvent.getPeriod() == NO_PERIOD) { //non-recurring
 
                     events.addEvent(newEvent);
-                    ui.eventAdded(newEvent, events.getNumEvents());
+                    ui.printEventAdded(newEvent, events.getNumEvents());
                     if (newEvent.getStartDate().getEventJavaDate().compareTo(currentDate.getTime()) < 0) {
                         ui.printEnteredEventOver();
                     }
@@ -455,14 +454,14 @@ public class Command {
                 }
 
             } catch (ClashException e) { //clash found
-                ui.scheduleClash(e.getClashEvent());
+                ui.printScheduleClash(e.getClashEvent());
             } catch (CostExceedsBudgetException e) { //budget exceeded in attempt to add concert
-                ui.costExceedsBudget(e.getConcert(), e.getBudget());
+                ui.printCostExceedsBudgetMsg(e.getConcert(), e.getBudget());
             } catch (StringIndexOutOfBoundsException | ArrayIndexOutOfBoundsException
                     | ParseException | NumberFormatException e) { //error interpreting info(wrong user input)
-                ui.newEntryFormatWrong();
+                ui.printNewEntryFormatWrong();
             } catch (EndBeforeStartException e) { //start time is after end time
-                ui.eventEndsBeforeStart();
+                ui.printEventEndsBeforeStart();
             }
         }
     }
@@ -506,21 +505,21 @@ public class Command {
      */
     private void addNewTodo(EventList events, UI ui) {
         if (continuation.isEmpty()) {
-            ui.eventDescriptionEmpty();
+            ui.printEventDescriptionEmpty();
             return;
         }
         try {
             EntryForToDo entryForToDo = new EntryForToDo().invoke(); //separate all info into relevant details
             Event newToDo = new ToDo(entryForToDo.getDescription(), entryForToDo.getDate());
             events.addNewTodo(newToDo);
-            ui.eventAdded(newToDo, events.getNumEvents());
+            ui.printEventAdded(newToDo, events.getNumEvents());
             Calendar currentDate = Calendar.getInstance();
             if (newToDo.getStartDate().getEventJavaDate().compareTo(currentDate.getTime()) < 0) {
                 ui.printEnteredEventOver();
             }
 
         } catch (StringIndexOutOfBoundsException | ArrayIndexOutOfBoundsException | ParseException e) {
-            ui.newEntryFormatWrong();
+            ui.printNewEntryFormatWrong();
         }
     }
 
@@ -532,9 +531,9 @@ public class Command {
             events.deleteEvent(eventNo - 1);
             ui.eventDeleted(currEvent);
         } catch (IndexOutOfBoundsException outOfBoundsE) {
-            ui.noSuchEvent();
+            ui.printNoSuchEvent();
         } catch (NumberFormatException notInteger) {
-            ui.notAnInteger();
+            ui.printNotAnInteger();
         }
     }
 
@@ -545,12 +544,12 @@ public class Command {
                 events.getEvent(eventNo - 1).markAsDone();
                 ui.eventDone(events.getEvent(eventNo - 1));
             } else {
-                ui.noSuchEvent();
+                ui.printNoSuchEvent();
             }
         } catch (IndexOutOfBoundsException outOfBoundsE) {
-            ui.noSuchEvent();
+            ui.printNoSuchEvent();
         } catch (NumberFormatException notInteger) {
-            ui.notAnInteger();
+            ui.printNotAnInteger();
         }
     }
 
@@ -591,17 +590,17 @@ public class Command {
             events.addEvent(newEvent);
             ui.rescheduleEvent(newEvent);
         } catch (ClashException clashE) {
-            ui.scheduleClash(clashE.getClashEvent());
+            ui.printScheduleClash(clashE.getClashEvent());
             newEvent.rescheduleStartDate(copyOfStartDate);
             newEvent.rescheduleEndDate(copyOfEndDate);
             events.undoDeletionOfEvent(newEvent);
         } catch (CostExceedsBudgetException e) {
-            ui.costExceedsBudget(e.getConcert(), e.getBudget());
+            ui.printCostExceedsBudgetMsg(e.getConcert(), e.getBudget());
             newEvent.rescheduleStartDate(copyOfStartDate);
             newEvent.rescheduleEndDate(copyOfEndDate);
             events.undoDeletionOfEvent(newEvent);
         } catch (Exception e) {
-            ui.eventEndsBeforeStart();
+            ui.printEventEndsBeforeStart();
             newEvent.rescheduleStartDate(copyOfStartDate);
             newEvent.rescheduleEndDate(copyOfEndDate);
             events.undoDeletionOfEvent(newEvent);
@@ -678,9 +677,9 @@ public class Command {
                 }
             }
         } catch (IndexOutOfBoundsException ne) {
-            ui.noSuchEvent();
+            ui.printNoSuchEvent();
         } catch (NumberFormatException numE) {
-            ui.notAnInteger();
+            ui.printNotAnInteger();
         }
     }
 
@@ -743,9 +742,9 @@ public class Command {
                 }
             }
         } catch (IndexOutOfBoundsException e) {
-            ui.noSuchEvent();
+            ui.printNoSuchEvent();
         } catch (NumberFormatException en) {
-            ui.notAnInteger();
+            ui.printNotAnInteger();
         }
     }
 
@@ -753,7 +752,7 @@ public class Command {
     public void manageInstruments(InstrumentList instruments, UI ui) {
         try {
             if (continuation.isEmpty()) {
-                ui.noSuchEvent();
+                ui.printNoSuchEvent();
                 return;
             }
             String splitInstrument[] = continuation.split("/");
@@ -790,7 +789,7 @@ public class Command {
                     break;
             }
         } catch (IndexOutOfBoundsException e) {
-            ui.noSuchEvent();
+            ui.printNoSuchEvent();
         }
     }
