+++ src/test/java/ListSortingAndClashDetectionTest.java
@@ -1,0 +1,113 @@
+//@@author Ryan-Wong-Ren-Wei
+
+import mistermusik.commons.events.eventtypes.Event;
+import mistermusik.commons.events.eventtypes.eventsubclasses.Concert;
+import mistermusik.commons.events.eventtypes.eventsubclasses.ToDo;
+import mistermusik.commons.events.eventtypes.eventsubclasses.recurringeventsubclasses.Lesson;
+import mistermusik.commons.events.eventtypes.eventsubclasses.recurringeventsubclasses.Practice;
+import mistermusik.logic.ClashException;
+import mistermusik.logic.EventList;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class ListSortingAndClashDetectionTest {
+    @Test
+    /**
+     * test clash handling for single event addition
+     */
+    public void clashTest() {
+        ArrayList<String> readFromFile = new ArrayList<String>();
+        String fileContent;
+        fileContent = "XT/fawpeifwe/02-12-2019";
+        readFromFile.add(fileContent);
+        fileContent = "XP/apiejfpwiefw/03-12-2019 1500/03-12-2019 1800";
+        readFromFile.add(fileContent);
+        fileContent = "XC/halloween/04-12-2019 1600/04-12-2019 1930/13";
+        readFromFile.add(fileContent);
+
+        EventList eventListTest = new EventList(readFromFile);
+        Event testEvent = new Practice("Horn practice", "3-12-2019 1400",
+                "3-12-2019 1600");
+        try {
+            eventListTest.addEvent(testEvent);
+        } catch (ClashException e) {
+            assertEquals(e.getClashEvent(), eventListTest.getEvent(1));
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    /**
+     * Test clash handling for recurring events
+     */
+    public void clashTestRecurring() {
+        ArrayList<String> readFromFile = new ArrayList<String>();
+        String fileContent;
+        fileContent = "XT/fawpeifwe/02-12-2019";
+        readFromFile.add(fileContent);
+        fileContent = "XP/apiejfpwiefw/03-12-2019 1500/03-12-2019 1800";
+        readFromFile.add(fileContent);
+        fileContent = "XC/halloween/04-12-2019 1600/04-12-2019 1930/3";
+        readFromFile.add(fileContent);
+
+        EventList eventListTest = new EventList(readFromFile);
+        Event testEvent = new Practice("Horn practice", "28-11-2019 1400",
+                "28-11-2019 1600");
+        try {
+            eventListTest.addRecurringEvent(testEvent, 4);
+        } catch (ClashException e) {
+            assertEquals(e.getClashEvent(), eventListTest.getEvent(1));
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    public void testSorting() throws Exception {
+        ArrayList<String> readFromFile = new ArrayList<String>();
+        String fileContent;
+        fileContent = "XT/fawpeifwe/02-12-2019";
+        readFromFile.add(fileContent);
+        fileContent = "XP/apiejfpwiefw/03-12-2019 1500/03-12-2019 1800";
+        readFromFile.add(fileContent);
+        fileContent = "XC/halloween/04-12-2019 1600/04-12-2019 1930/5";
+        readFromFile.add(fileContent);
+
+        EventList eventListTest = new EventList(readFromFile);
+        boolean succeeded = true;
+        Event testEvent1 = new Practice("Horn practice", "05-12-2019 1400",
+                "05-12-2019 1600");
+        Event testEvent2 = new Lesson("Full Orchestra rehearsal", "03-12-2019 1400",
+                "03-12-2019 1500");
+        Event testEvent3 = new ToDo("Complete theory homework CS2113", "01-12-2019");
+
+        eventListTest.addEvent(testEvent1);
+        eventListTest.addEvent(testEvent2);
+        eventListTest.addNewTodo(testEvent3);
+        eventListTest.sortList();
+        ArrayList<Event> eventListCompare = new ArrayList<>();
+
+        eventListCompare.add(new ToDo("Complete theory homework CS2113", "01-12-2019"));
+        eventListCompare.add(new ToDo("fawpeifwe", "02-12-2019"));
+        eventListCompare.add(new Lesson("Full Orchestra rehearsal", "03-12-2019 1400", "03-12-2019 1500"));
+        eventListCompare.add(new Practice("apiejfpwiefw", "03-12-2019 1500", "03-12-2019 1800"));
+        eventListCompare.add(new Concert("halloween", "04-12-2019 1600", "04-12-2019 1930", 5));
+        eventListCompare.add(new Practice("Horn practice", "05-12-2019 1400", "05-12-2019 1600"));
+
+        int i = 0;
+        for (Event currEvent : eventListTest.getEventArrayList()) {
+//            System.out.println(currEvent.toString());
+//            System.out.println(eventListCompare.get(i).toString());
+            if (!currEvent.toString().equals(eventListCompare.get(i).toString())) {
+                succeeded = false;
+            }
+            ++i;
+        }
+
+        assertEquals(true, succeeded);
+    }
+}
